fix: prevent compressible account funding with 1 epoch (#1993)
* fix: prevent compressible account funding with 1 epoch

* removed results from toplevel test functions

* add ctoken create account tests

* add close account tests

* test: add ctoken transfer tests

* add create ata tests

* restore program ci

* add compress and close tests

* fix: spl instruction compatibility

* add rent constant

* fix failing tests asserts

* add mint duplicates check

* compressible add tests and overflow guards

* refactor: use array map and tinyvec instead of arrayVec

* refactor: unify output compressed indices into one

* refactor: ctoken instruction discriminators

+++ sdk-libs/compressed-token-sdk/src/instructions/create_token_account/instruction.rs
@@ -26,7 +26,7 @@ pub struct CreateCompressibleTokenAccount {
     /// The rent recipient PDA (fee_payer_pda in processor)
     pub rent_sponsor: Pubkey,
     /// Number of epochs of rent to prepay
-    pub pre_pay_num_epochs: u64,
+    pub pre_pay_num_epochs: u8,
     /// Initial lamports to top up for rent payments (optional)
     pub lamports_per_write: Option<u32>,
     pub compress_to_account_pubkey: Option<CompressToPubkey>,
